refactor: split the three files over the 500-LoC cap
Unblocks the failing `File size cap` check on #754. Split rather than
allow-listed: the allow-list wants a rationale and a tracking issue, and
each of these has a real seam.

- `ops/moe_grouped_a.rs` 538 -> 448. The three `moe_topk_*` routers move
  to `moe_grouped_a/topk.rs`, along the banner the file already drew:
  they answer "which experts", the rest is the GEMM that runs once that
  is known. They differ only in the normaliser the router was trained
  with, and picking the wrong one is silent, so they read better side by
  side.

- `qsa.rs` 504 -> 440. `snapshot_aux`/`restore_aux` move to
  `qsa_snapshot.rs`, following this file's own `#[path]` sibling idiom.
  Everything left computes a selection; these two preserve and restore
  the state computing it consumed — the thing speculative decoding gets
  wrong.

- `moe/mod.rs` 502 -> 435. The four expert-addressing types move to
  `moe/tables.rs`. They are the layer's vocabulary, not its behaviour.

Three things the compiler caught, all of them the split's own doing:

- `#[path]` inheritance. `ops.rs` loads `moe_grouped_a.rs` with an
  explicit path, so a plain `mod topk;` inside it resolves against
  `ops/`, not a `moe_grouped_a/` subdirectory. The child gets an
  explicit `#[path]` too, matching how the parent is loaded.
- `Bf16SharedExpert`'s fields and `new` were module-private and are
  reached from four sibling files. They are `pub(super)` now: the split
  must not narrow what was already visible.
- Two imports the moved code turned out not to use.

`cargo check --workspace` clean, rustfmt clean.
